chore: rename edits made to C code
