+++ src/c/compression_zstd/clinic/decompressor.c.h
@@ -2,7 +2,7 @@
 preserve
 [clinic start generated code]*/
 
-#include "backports_zstd_redef.h"
+#include "backports_zstd_edits.h"
 
 PyDoc_STRVAR(_zstd_ZstdDecompressor___init____doc__,
 "ZstdDecompressor(zstd_dict=None, options=None)\n"
@@ -60,7 +60,7 @@ _zstd_ZstdDecompressor___init__(PyObject *self, PyObject *args, PyObject *kwargs
     PyObject *zstd_dict = Py_None;
     PyObject *options = Py_None;
 
-    fastargs = _backportszstdredef__PyArg_UnpackKeywords(_PyTuple_CAST(args)->ob_item, nargs, kwargs, NULL, &_parser,
+    fastargs = BACKPORTSZSTD__PyArg_UnpackKeywords(_PyTuple_CAST(args)->ob_item, nargs, kwargs, NULL, &_parser,
             /*minpos*/ 0, /*maxpos*/ 2, /*minkw*/ 0, /*varpos*/ 0, argsbuf);
     if (!fastargs) {
         goto exit;
@@ -187,7 +187,7 @@ _zstd_ZstdDecompressor_decompress(PyObject *self, PyObject *const *args, Py_ssiz
     Py_buffer data = {NULL, NULL};
     Py_ssize_t max_length = -1;
 
-    args = _backportszstdredef__PyArg_UnpackKeywords(args, nargs, NULL, kwnames, &_parser,
+    args = BACKPORTSZSTD__PyArg_UnpackKeywords(args, nargs, NULL, kwnames, &_parser,
             /*minpos*/ 1, /*maxpos*/ 2, /*minkw*/ 0, /*varpos*/ 0, argsbuf);
     if (!args) {
         goto exit;
@@ -200,7 +200,7 @@ _zstd_ZstdDecompressor_decompress(PyObject *self, PyObject *const *args, Py_ssiz
     }
     {
         Py_ssize_t ival = -1;
-        PyObject *iobj = _backportszstdredef__PyNumber_Index(args[1]);
+        PyObject *iobj = BACKPORTSZSTD__PyNumber_Index(args[1]);
         if (iobj != NULL) {
             ival = PyLong_AsSsize_t(iobj);
             Py_DECREF(iobj);
